CurrentStation, Exit Vehicle, Debug and Drag
Current station removal on player is simplified (so it works), but may not be robust.
Finding current station fixed. Before it wasn't checking parents, and the collider was on the TileMap.

Exiting a vehicle on a station has been fixed. The player wasn't able to determine if it was in a station because it's rigidbody wasn't simulated. Now the vehicle has "currentstation" and the player checks that

Debug checks added to annoying debug statements that are often used.

Drag implemented for entering/exiting stations on entities.
Still need to implement ship locking when exiting the vehicle.